\name{genotypes}
\alias{genotypes}
\docType{data}
\title{
Genotypes for 100 SNPs and 75 individuals
}

\description{
A data frame with 75 rows (individuals) and 100 columns (SNPs). The entries of the genotype matrix are 0, 1 and 2. There are no missing values. 
}

\usage{data(genotypes)}

\seealso{ \code{\link{SNPs}}}

\examples{
data(genotypes)
}

\keyword{datasets}
