\name{PAGWAS-package}
\alias{PAGWAS-package}
\alias{PAGWAS}
\docType{package}
\title{
Pathway Analysis Methods for Genomewide Association Data}

\description{Bayesian hierarchical methods for pathway analysis of genomewide association data: Normal/Bayes factors and Sparse Normal/Adaptive lasso. The Frequentist Fisher's method is included as well.}

\details{
\tabular{ll}{
Package: \tab PAGWAS\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2015-12-02\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}

}

\author{
Marina Evangelou
Maintainer: Marina Evangelou <m.evangelou@ic.ac.ukt>
}

\references{
Evangelou, M., Dudbridge, F., Wernisch, L. (2014). Two novel pathway analysis methods based on a hierarchical model. Bioinformatics 30(5), 690 - 697

Evangelou, M., Rendon, A., Ouhewand, W. H., Wernisch, L., Dudbridge, F. (2012) Comparison of methods for competitive tests of pathway analysis. Plos One 7(7): e41018
}

\keyword{ package }
