\name{pathways}
\alias{pathways}
\docType{data}
\title{
A list of 2 pathways with their gene members
}

\description{A list of 2 pathways with their gene members}

\usage{data(pathways)
}

\seealso{
\code{\link{genes}}, \code{\link{SNPs}}
}

\examples{
data(pathways)
pathways[[1]]
}

\keyword{datasets}