\name{PAGWAS-package}
\alias{PAGWAS-package}
\alias{PAGWAS}
\docType{package}
\title{
Pathway analysis methods for genomewide association data
}

\description{
The two Bayesian hierarchical pathway analysis methods Sparse normal/Adaptive lasso and Normal/Bayes factors proposed by Evangelou et al (2013) are implemented. Further, the well-known Fisher's method is implemented
}

\details{
\tabular{ll}{
Package: \tab PAGWAS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-15\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}

}

\author{
Marina Evangelou  <marina.evangelou@cimr.cam.ac.uk>
}

\references{
Evangelou, M., Dudbridge, F., Wernisch, L. (2013). Two novel pathway analysis methods based on a hierarchical model. Bioinformatics (to appear)

Evangelou, M., Rendon, A., Ouhewand, W. H., Wernisch, L., Dudbridge, F. (2012). Comparison of methods for competitive tests of pathway analysis. PLoS ONE 7(7), e41018
}

\keyword{ package }
