\name{pathways}
\alias{pathways}
\docType{data}
\title{
A list of 2 pathways with their gene members
}

\description{A list of two pathways. The gene members of each pathway are given.}

\usage{data(pathways)}

\seealso{
\code{\link{genes}}, \code{\link{SNPs}}
}

\examples{
data(pathways)
}

\keyword{datasets}